move preview related to folder
